requirements.txt
pillow
sudokum